+++ src/main/kotlin/icu/windea/pls/cwt/config/CwtConfigType.kt
@@ -3,6 +3,9 @@ package icu.windea.pls.cwt.config
 enum class CwtConfigType(val text:String){
 	Type("type"),
 	Subtype("subtype"),
+	Value("value"),
+	ValueValue("value value"),
 	Enum("enum"),
-	EnumValue("enum value")
+	EnumValue("enum value"),
+	Alias("alias")
 }