+++ src/main/kotlin/icu/windea/pls/localisation/codeInsight/ParadoxLocalisationCompletionContributor.kt
@@ -7,33 +7,24 @@ import com.intellij.codeInsight.lookup.*
 import com.intellij.patterns.PlatformPatterns.*
 import com.intellij.util.*
 import icu.windea.pls.*
+import icu.windea.pls.localisation.psi.*
 import icu.windea.pls.localisation.psi.ParadoxLocalisationTypes.*
 
 class ParadoxLocalisationCompletionContributor : CompletionContributor() {
 	companion object{
 		private val localePattern = psiElement(LOCALE_ID)
-		private val sequentialNumberPattern = psiElement(SEQUENTIAL_NUMBER_ID)
-		private val colorIdPattern = psiElement(COLOR_ID)
-		//private val commandScopePattern = psiElement(COMMAND_SCOPE_ID)
-		//private val commandFieldPattern = psiElement(COMMAND_FIELD_ID)
+		//private val sequentialNumberPattern = psiElement(SEQUENTIAL_NUMBER_ID)
+		//private val colorIdPattern = psiElement(COLOR_ID)
+		private val commandFieldPattern = psiElement(COMMAND_FIELD_ID)
 		
 		private val localeElements = getConfig().locales.map {
 			LookupElementBuilder.create(it.name).withIcon(it.icon).withTailText(it.tailText,true)
 		}
-		private val sequentialNumberElements = getConfig().sequentialNumbers.map {
-			LookupElementBuilder.create(it.name).withIcon(it.icon).withTailText(it.tailText,true)
-		}
-		private val colorElements = getConfig().colors.map{
-			LookupElementBuilder.create(it.name).withIcon(it.icon).withTailText(it.tailText,true)
-		}
-		//private val primaryCommandScopeElements = config.primaryCommandScopes.map {
-		//	LookupElementBuilder.create(it.name).withTypeText(it.description).withIcon(localisationCommandScopeIcon)
+		//private val sequentialNumberElements = getConfig().sequentialNumbers.map {
+		//	LookupElementBuilder.create(it.name).withIcon(it.icon).withTailText(it.tailText,true)
 		//}
-		//private val secondaryCommandScopeElements = config.secondaryCommandScopes.map {
-		//	LookupElementBuilder.create(it.name).withTypeText(it.description).withIcon(localisationCommandScopeIcon)
-		//}
-		//private val commandFieldElements = config.commandFields.map {
-		//	LookupElementBuilder.create(it.name).withTypeText(it.description).withIcon(localisationCommandFieldIcon)
+		//private val colorElements = getConfig().colors.map{
+		//	LookupElementBuilder.create(it.name).withIcon(it.icon).withTailText(it.tailText,true)
 		//}
 	}
 	
@@ -43,50 +34,34 @@ class ParadoxLocalisationCompletionContributor : CompletionContributor() {
 		}
 	}
 	
-	class SequentialNumberCompletionProvider : CompletionProvider<CompletionParameters>() {
-		override fun addCompletions(parameters: CompletionParameters, context: ProcessingContext, result: CompletionResultSet) {
-			result.addAllElements(sequentialNumberElements)
-		}
-	}
-	
-	class ColorCompletionProvider : CompletionProvider<CompletionParameters>() {
-		override fun addCompletions(parameters: CompletionParameters, context: ProcessingContext, result: CompletionResultSet) {
-			result.addAllElements(colorElements)
-		}
-	}
-	
-	//class CommandCompletionProvider : CompletionProvider<CompletionParameters>() {
+	//class SequentialNumberCompletionProvider : CompletionProvider<CompletionParameters>() {
 	//	override fun addCompletions(parameters: CompletionParameters, context: ProcessingContext, result: CompletionResultSet) {
-	//		val position  = parameters.position
-	//		val prefix = position.text.dropLast(dummyIdentifierLength).trim()
-	//		if(prefix == eventTargetPrefix) return 
-	//		
-	//		val parent = position.parent //COMMAND_SCOPE, COMMAND_FIELD
-	//		if(parent !is ParadoxLocalisationCommandIdentifier) return
-	//		
-	//		val prev = parent.prevIdentifier
-	//		if(prev == null){
-	//			//primaryCommandScope, secondaryCommandScope, event_target
-	//			result.addAllElements(primaryCommandScopeElements)
-	//			result.addAllElements(secondaryCommandScopeElements)
-	//		}else{
-	//			//secondaryCommandScope
-	//			result.addAllElements(secondaryCommandScopeElements)
-	//		}
-	//		val next = parent.nextIdentifier
-	//		if(next == null){
-	//			//commandField, scopeVariable, scriptedLoc
-	//			result.addAllElements(commandFieldElements)
-	//		}
+	//		result.addAllElements(sequentialNumberElements)
+	//	}
+	//}
+	//
+	//class ColorCompletionProvider : CompletionProvider<CompletionParameters>() {
+	//	override fun addCompletions(parameters: CompletionParameters, context: ProcessingContext, result: CompletionResultSet) {
+	//		result.addAllElements(colorElements)
 	//	}
 	//}
 	
+	class CommandFieldCompletionProvider : CompletionProvider<CompletionParameters>() {
+		override fun addCompletions(parameters: CompletionParameters, context: ProcessingContext, result: CompletionResultSet) {
+			val position = parameters.position //COMMAND_FIELD_ID
+			val commandField = position.parent
+			val project = position.project
+			val gameType = parameters.originalFile.paradoxFileInfo?.gameType?:return
+			val configGroup = getConfig(project).get(gameType)?:return
+			completeLocalisationCommand(commandField,configGroup, result)
+		}
+	}
+	
 	init {
 		extend(CompletionType.BASIC, localePattern, LocaleCompletionProvider())
-		extend(CompletionType.BASIC, sequentialNumberPattern, SequentialNumberCompletionProvider()) //无法被匹配，但仍然留着
-		extend(CompletionType.BASIC, colorIdPattern, ColorCompletionProvider()) //无法被匹配，但仍然留着
-		//extend(CompletionType.BASIC, commandScopePattern, CommandCompletionProvider())
-		//extend(CompletionType.BASIC, commandFieldPattern, CommandCompletionProvider())
+		//extend(CompletionType.BASIC, sequentialNumberPattern, SequentialNumberCompletionProvider()) //NOTE 无法被匹配
+		//extend(CompletionType.BASIC, colorIdPattern, ColorCompletionProvider()) //NOTE 无法被匹配
+		extend(CompletionType.BASIC, commandFieldPattern, CommandFieldCompletionProvider()) //TODO 匹配scope
 	}
 	
 	override fun beforeCompletion(context: CompletionInitializationContext) {